Oganize Adapters
